requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
jinja2==3.1.4
python-multipart==0.0.9

# Auth
python-jose[cryptography]==3.3.0
bcrypt==4.2.0

# Database
aiosqlite==0.20.0
sqlalchemy==2.0.35

# Video download (Douyin, Xiaohongshu, 1688)
yt-dlp>=2024.12.0

# Speech-to-text / Translation
openai>=1.50.0

# Browser automation (Douyin/Xiaohongshu 접근)
playwright==1.49.0

# Video processing / Subtitle
pysrt==1.1.2

# Encryption
cryptography==43.0.0

# Utils
python-dotenv==1.0.1
httpx==0.27.2
aiofiles==24.1.0
pydantic==2.9.0
pydantic-settings==2.5.0
